Rename command for asset that will handle  soft_deletion, restore and destroy.
