--- app/commands/assets/soft_delete_with_relations.rb
@@ -1,19 +1,0 @@
-module Assets
-  class SoftDeleteWithRelations
-    attr_reader :asset
-
-    def initialize(asset)
-      @asset = asset
-    end
-
-    def execute
-      time = Time.now
-      # first update playlists
-      asset.playlists.update_all(['tracks_count = tracks_count - 1, playlists.updated_at = ?', Time.now]) unless asset.playlists.empty?
-      asset.comments.update_all(deleted_at: time)
-      asset.tracks.update_all(deleted_at: time)
-      asset.listens.update_all(deleted_at: time)
-      asset.soft_delete
-    end
-  end
-end
